New class: DataHandler
The purpose of this class is to abstract the process of managing data. Notebooks
are represented by a JSONObject of the form:
	{"title":"exampleTitle", "description":"exampleDescription"}.
All notebooks are contained in a JSONArray. This array is then converted into a String.

Up to now, the main method of the class is getNotebooks(). It opens the file where
the JSON String described above is saved and creates the Array<Notebook> needed by the
RecyclerView Adapter to display the notebooks.

JSONObjects were chosen after a little researching about data representation.
It is easier to handle JSONbjects than plain text with separators or another self-made
implementation.

To write a String in a File or to read a file as a String, Google Guava library was
used because of its practicality.