Data is managed through a singleton class
Now, the DataHandler class is wrapped by a Singleton class called Data.
Doing so allows to create an unique DataHandler object and to use it in different
Activities. This feature is needed because it will be necesary to create an activity
in which the user could add new notebooks. That implies adding new information to the
DataHandler object.