Bug fixed
Data was been loaded multiple times to the RecyclerView when the user pressed the
"Back" button and then went back to the app (Also when the app was stopped after a
time and then killed by the operating system due to system constraints). This happened
because OnBackPressed() destroys the app stopping it first, and when the activity begins
to stop, the system calls OnSaveInstanceState()to save information about each view
object of the activity layout in a Bundle object. When the activity is recreated the views
are restored and in the previous implementation of getNotebooks(), data was loaded from the
file without checking if it was done before.