Floating button implemented
Small changes were made to support this new feature. Now the RecyclerView which
contains the notebooks is added from the xml file. Both the RecyclerView and the
FloatingButton are contained in a CoordinatorLayout. This allows to hover the
FloatingButton over the RecyclerView.

Some other interesting changes in MainActivity.java:

- Use of Preconditions.checkArgument(...); to ensure that the value returned
from findViewById() is not null.

- Use of string resources instead of harcoded strings.

- rvNotebooks.setVisibility(View.INVISIBLE); is used to hide
the RecyclerView in order to be able to show the message in case that there are no
notebooks to display.