Change in README.md (App overview)
Originally:
"... Once inside a notebook, we will see the different notes of the user. They will
be able to be displayed as a grid or a list. By default they will be sorted by
order of creation, but it will be possible to sort them in lexicographical order. ..."

I think that the lexicographical order is not a useful feature. Instead, It would be
useful if the user could sort notes in a relevant order for him (manually, moving
each note after long-pressing it).